fix: handle negative input intervals; add tests
